packages/policy-asset-fetcher/requirements.txt
pymongo==4.3.3
python-dotenv==1.0.0
httplib2==0.22.0
blockfrost-python==0.5.3